chore: Update dependencies and do a small refactor
This commit updates all the Go dependencies to their latest versions,
updates workflow files a bit and does a small refactor, such as fixing
typos and following best practices.

grpc.DialContext still needs to be updated, but I don't have enough
experience with it or with this tool to be confident that I properly
used it (due to the changes in the name resolver or the state changes).